% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggthemeassist-package.R, R/ggthemeassist.R
\docType{package}
\name{ggthemeassist}
\alias{ggthemeassist}
\alias{ggthemeassist-package}
\title{RStudio Addins}
\usage{
ggthemeassist()
}
\value{
\code{ggthemeassist} returns a character vector.
}
\description{
Addin to help style your ggplot2 themes

\code{ggthemeassist} is a RStudio-Addin that delivers a graphical interface for editing ggplot2 theme elements.
}
\details{
To run the addin, highlight a ggplot2-object in your current script and select \code{ggthemeassist} from the Addins-menu within RStudio. After editing themes and terminating the addin, a character string containing the desired changes is inserted in your current script.
}

