% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggThemeAssist-package.R, R/ggThemeAssist.R
\docType{package}
\name{ggThemeAssist}
\alias{ggThemeAssist}
\alias{ggThemeAssist-package}
\title{RStudio Addins}
\usage{
ggThemeAssist()
}
\value{
\code{ggThemeAssist} returns a character vector.
}
\description{
Addin to help style your ggplot2 themes

\code{ggThemeAssist} is a RStudio-Addin that delivers a graphical interface for editing ggplot2 theme elements.
}
\details{
To run the addin, highlight a ggplot2-object in your current script and select \code{ggThemeAssist} from the Addins-menu within RStudio. After editing themes and terminating the addin, a character string containing the desired changes is inserted in your current script.
}

