% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggThemeAssist-package.R, R/ggThemeAssist.R
\docType{package}
\name{ggThemeAssist}
\alias{ggThemeAssist}
\alias{ggThemeAssist-package}
\title{RStudio Addins}
\usage{
ggThemeAssist(text = "")
}
\arguments{
\item{text}{(string) The name of a ggplot2 object, or code to create a ggplot2 object.  Only for use in the console; has no effect when running as an RStudio addin.}
}
\value{
\code{ggThemeAssist} returns a character vector.
}
\description{
Addin to help style your ggplot2 themes

\code{ggThemeAssist} is a RStudio-Addin that delivers a graphical interface for editing ggplot2 theme elements.
}
\details{
To run as an RStudio addin, highlight a ggplot2 object in your current script and select \code{ggThemeAssist} from the Addins-menu within RStudio. After editing themes and terminating the addin, a character string containing the desired changes is inserted in your current script.

To run as a Shiny Gadget outside of RStudio, pass the name of the ggplot2-objeobject to the `text` parameter of this function.  The gadget will be opened in the default browser.  After editing themes and terminating the addin, a character string containing the desired changes will be echoed on thethe command line.
}

